requirements.txt
selenium
webdriver_manager
allure-pytest
faker
factory_boy